+++ CHANGELOG.rst
@@ -1,3 +1,13 @@
+Version 9.0.0
+-------------
+
+**Removed**:
+
+- The ``loop`` method for the ``Schema``, which was considered deprecated and unsafe, has been removed.
+- The ``iterations`` parameter for all methods of ``Schema`` has been removed. Instead, you now have to specify the number of iterations on instantiation of ``Schema`` passing the ``i`` parameter.
+- The ``iterator`` method for ``Schema`` has been removed. Instead, you can now use an instance of ``Schema`` directly as an iterator.
+
+
 Version 8.0.0
 -------------